ui/requirements.txt
streamlit==1.24.0
requests==2.31.0
qrcode==7.4.2
python-dotenv==1.0.0
Pillow==9.5.0
